conda_requirements.txt
bioawk=1.0
biopython
blast=2.7.1
boto3=1.7.1
bowtie2
bwa=0.7.17
bzip2=1.0.6
ec2-metadata
ete3
gatk=3.8
gatk4=4.0.4.0
ipython
jellyfish=1.1.11
jupyter
matplotlib
krona=2.7
numpy
pandas
picard
prinseq=0.20.4
psutil
pyfaidx
pysam
pyyaml
python=3.7
samtools
scikit-bio
scikit-learn
scipy
seqtk=1.3
snakemake=5.2.2
wgsim=1.0
